src/data_generation/consumer/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.4